Completes task 382 which wires the CardUI to subscribe to CardDrawnEvents
