--- Assets/Scripts/UI/Cards/CardPopupPanel.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI; 
-
-public class CardPopupUI : MonoBehaviour
-{
-    [Header("UI References")]
-    [SerializeField] private Text cardTitleText;
-    [SerializeField] private Text cardBodyText;
-    [SerializeField] private Button okButton;
-    [SerializeField] private CanvasGroup CanvasGroup;
-    [SerializeField] private Image artworkImage;
-
-    //More fields will be added later on 
-}
